# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An `API` for cross-platform custom orchestration of execution steps\nwithout any third-party dependencies. Based on `DAG`, it implements the scheduling\nfunction of sequential execution of dependent steps and concurrent execution of\nnon-dependent steps. \u003cbr /\u003e\u003cbr /\u003e It provides `API` remote operation mode, batch\nexecution of `Shell` , `Powershell` , `Python` and other commands, and easily\ncompletes common management tasks such as running automated operation and maintenance\nscripts, polling processes, installing or uninstalling software, updating applications,\nand installing patches.",
        "title": "DAG Flow",
        "contact": {
            "name": "dagflow",
            "url": "https://github.com/busyster996/dagflow/issues"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/busyster996/dagflow/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/task": {
            "get": {
                "description": "获取任务列表, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "分页大小",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_STaskListDetailRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "创建任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "description": "任务内容",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.STaskReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_STaskCreateRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/task/{task}/step": {
            "get": {
                "description": "获取指定任务的步骤列表, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "步骤"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SStepsRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/task/{task}/step/{step}": {
            "get": {
                "description": "获取步骤详情, 支持SSE订阅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "步骤"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "步骤名称",
                        "name": "step",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SStepRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "put": {
                "description": "管理指定任务的指定步骤, 支持暂停、恢复、终止、超时暂停自动恢复",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "步骤"
                ],
                "summary": "管理",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "步骤名称",
                        "name": "step",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "paused",
                            "kill",
                            "pause",
                            "resume"
                        ],
                        "type": "string",
                        "default": "paused",
                        "description": "操作项",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "1m",
                        "description": "暂停多久, 如果没设置则需要手工恢复",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/task/{task}/step/{step}/log": {
            "get": {
                "description": "指定任务指定步骤的执行输出, 支持WS长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "步骤"
                ],
                "summary": "日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "步骤名称",
                        "name": "step",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SStepLogsRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
        "/api/v1/task/{task}/workspace": {
            "get": {
                "description": "获取目录列表或下载指定文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作目录"
                ],
                "summary": "列表或下载",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-types_SFileListRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "post": {
                "description": "上传文件或目录",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作目录"
                ],
                "summary": "上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定目录或文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作目录"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.SBase-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        },
        "types.Code": {
            "type": "integer",
            "enum": [
                0,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeRunning",
                "CodeFailed",
                "CodeNoData",
                "CodePending",
                "CodePaused",
                "CodeSkipped"
            ]
        },
        "types.SBase-any": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SBase-types_SFileListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "$ref": "#/definitions/types.SFileListRes"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "types.SBase-types_SStepLogsRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SStepLogRes"
                    }
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SBase-types_SStepRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "$ref": "#/definitions/types.SStepRes"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SBase-types_SStepsRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SStepRes"
                    }
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SBase-types_STaskCreateRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "$ref": "#/definitions/types.STaskCreateRes"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SBase-types_STaskListDetailRes": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.Code"
                },
                "data": {
                    "$ref": "#/definitions/types.STaskListDetailRes"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "types.SEnv": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "types.SFileListRes": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SFileRes"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.SPageRes": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.SStepLogRes": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "line": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "types.SStepReq": {
            "type": "object",
            "required": [
                "content",
                "type"
            ],
            "properties": {
                "action": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "depends": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "desc": {
                    "type": "string"
                },
                "disable": {
                    "type": "boolean"
                },
                "env": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SEnv"
                    }
                },
                "name": {
                    "type": "string"
                },
                "retryPolicy": {
                    "$ref": "#/definitions/types.SRetryPolicy"
                },
                "rule": {
                    "type": "string"
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "types.SStepRes": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "depends": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "desc": {
                    "type": "string"
                },
                "disable": {
                    "type": "boolean"
                },
                "env": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SEnv"
                    }
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "retryPolicy": {
                    "$ref": "#/definitions/types.SRetryPolicy"
                },
                "rule": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "time": {
                    "$ref": "#/definitions/types.STimeRes"
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "types.STaskCreateRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.STaskListDetailRes": {
            "type": "object",
            "properties": {
                "page": {
                    "$ref": "#/definitions/types.SPageRes"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.STaskRes"
                    }
                }
            }
        },
        "types.STaskReq": {
            "type": "object",
            "required": [
                "step"
            ],
            "properties": {
                "delayed": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "disable": {
                    "type": "boolean"
                },
                "env": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SEnv"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "step": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SStepReq"
                    }
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
        "types.STaskRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "disable": {
                    "type": "boolean"
                },
                "env": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SEnv"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "time": {
                    "$ref": "#/definitions/types.STimeRes"
                },
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
...
    }
}