# server/api/v1/task/post.go
package task

import (
	"github.com/busyster996/dagflow/pkg/logx"
	"github.com/gin-gonic/gin"

	"github.com/busyster996/dagflow/server/api/base"
	"github.com/busyster996/dagflow/server/service"
	"github.com/busyster996/dagflow/server/types"
)

// Post
// @Summary		创建
// @Description	创建任务
// @Tags		任务
// @Accept		application/json
// @Produce		application/json
// @Param		task body types.STaskReq true "任务内容"
// @Success		200 {object} types.SBase[types.STaskCreateRes]
// @Failure		500 {object} types.SBase[any]
// @Router		/api/v1/task [post]
func Post(c *gin.Context) {
	var req = new(types.STaskReq)
	if err := c.ShouldBind(req); err != nil {
		logx.Errorln(err)
		base.Send(c, base.WithCode[any](types.CodeFailed).WithError(err))
		return
	}

	if err := service.Task(req.Name).Create(req); err != nil {
		base.Send(c, base.WithCode[any](types.CodeFailed).WithError(err))
		return
	}

	c.Request.Header.Set(types.XTaskName, req.Name)
	c.Header(types.XTaskName, req.Name)

	base.Send(c, base.WithData(&types.STaskCreateRes{
		Name:  req.Name,
		Count: int64(len(req.Step)),
	}))
}